+++ ProyectosSDL/HolaSDL/Scoremarker.h
@@ -9,7 +9,7 @@ private:
   int _score;
 
 public:
-  ScoreMarker(float x, float y, int width, int height, SDL_Color color, Font *font);
+  ScoreMarker(float32 x, float32 y, float32 width, float32 height, SDL_Color color, Font *font);
   ~ScoreMarker(){};
   void update();
   virtual std::istream &deserialize(std::istream &out);
