+++ ProyectosSDL/HolaSDL/Text.cpp
@@ -3,7 +3,7 @@
 /**
  * Constructors.
  */
-Text::Text(Font *font, float x, float y, int width, int height, const SDL_Color &color, const string &text)
+Text::Text(Font *font, float32 x, float32 y, float32 width, float32 height, const SDL_Color &color, const string &text)
     : _font(font), GameObject(x, y, width, height), _color(color), _text(text)
 {
