+++ ProyectosSDL/HolaSDL/Text.h
@@ -15,7 +15,7 @@ private:
   int _width, _height;
 
 public:
-  Text(Font *font, float x, float y, int width, int height, const SDL_Color &color, const string &text);
+  Text(Font *font, float32 x, float32 y, float32 width, float32 height, const SDL_Color &color, const string &text);
   ~Text();
   void setText(const string newText);
   SDL_Rect getRect() const;
