+++ ProyectosSDL/HolaSDL/Timer.h
@@ -9,7 +9,7 @@ private:
   int time;
 
 public:
-  Timer(float x, float y, int width, int height, SDL_Color color, Font *font);
+  Timer(float32 x, float32 y, float32 width, float32 height, SDL_Color color, Font *font);
   ~Timer(){};
   void update();
   virtual std::istream &deserialize(std::istream &out);
